Refactor helpers
Naming them by function rather than controller and thereby removing a
lot of empty file clutter. Singular names to hint that these are not
controllers.
