--- app/helpers/about_helper.rb
@@ -1,2 +1,0 @@
-module AboutHelper
-end
