+++ app/helpers/application_helper.rb
@@ -1,69 +1,6 @@
+##
+# Miscellaneous helper methods
 module ApplicationHelper
-
-  def xr_icon(model, options = {}, html_options = {})
-    icon = model.icon
-
-    if options.has_key?(:light)
-      icon = icon.split(".")
-      icon = icon[0] + "-light." + icon[1]
-    end
-
-    unless html_options.has_key?(:alt)
-      html_options[:alt] = model.label
-    end
-
-    # Set intrinsic dimensions (actual display dimensions assumed to be set by CSS!)
-    html_options[:width] = 72
-    html_options[:height] = 72
-
-    image_tag icon, html_options
-  end
-
-  # Bootstrap Icons (icon font method)
-  def bs_icon(icon, options = {})
-    classes = ["bi-#{icon}"].append(options.delete(:class)).compact
-    content_tag :i, "", options.merge(class: classes)
-  end
-
-  def create_icon(options = {})
-    bs_icon "plus", options
-  end
-
-  def create_inline_icon(options = {})
-    bs_icon "plus-square", options
-  end
-
-  def edit_icon(options = {})
-    bs_icon "pencil", options
-  end
-
-  def edit_inline_icon(options = {})
-    bs_icon "pencil-square", options
-  end
-
-  def delete_icon(options = {})
-    bs_icon "x-lg", options
-  end
-
-  def delete_inline_icon(options = {})
-    bs_icon "x-square", options
-  end
-
-  def confirm_icon(options = {})
-    bs_icon "check-lg", options
-  end
-
-  def confirm_inline_icon(options = {})
-    bs_icon "check-square", options
-  end
-
-  def active_class(path)
-    "active" if current_page?(path)
-  end
-
-  def active_aria(path)
-    'aria-current="page"'.html_safe if current_page?(path)
-  end
 
   def na_value
     '<abbr title="Unknown or missing value" class="initialism text-muted">NA</abbr>'.html_safe
