--- app/helpers/articles_helper.rb
@@ -1,2 +1,0 @@
-module ArticlesHelper
-end
