--- app/helpers/c14_labs_helper.rb
@@ -1,2 +1,0 @@
-module C14LabsHelper
-end
