+++ app/helpers/cite_as_helper.rb
@@ -1,5 +1,9 @@
+##
 # Helper methods for generating citations to XRONOS pages
 module CiteAsHelper
+
+  ##
+  # Generate citation for an article
   def article_citation(article)
     <<~TEXT
       #{article.author_name} (#{article.publication_date}), #{article.title}.
@@ -9,7 +13,10 @@ module CiteAsHelper
     TEXT
   end
 
+  private
+
   def article_citation_publication(article)
     article.news_section? ? "XRONOS News" : "XRONOS"
   end
+
 end
