--- app/helpers/curate_helper.rb
@@ -1,2 +1,0 @@
-module CurateHelper
-end
