--- app/helpers/materials_helper.rb
@@ -1,2 +1,0 @@
-module MaterialsHelper
-end
