--- app/helpers/measurement_states_helper.rb
@@ -1,2 +1,0 @@
-module MeasurementStatesHelper
-end
