--- app/helpers/meta_helper.rb
@@ -1,5 +1,0 @@
-module MetaHelper
-  def meta(name, default = nil)
-    content_for?(name) ? content_for(name) : default
-  end
-end
