--- app/helpers/references_helper.rb
@@ -1,3 +1,0 @@
-module ReferencesHelper
-  include DataTableHelper
-end
