--- app/helpers/samples_helper.rb
@@ -1,2 +1,0 @@
-module SamplesHelper
-end
