+++ app/helpers/search_helper.rb
@@ -1,4 +1,6 @@
-module SearchesHelper
+##
+# Helper methods for search
+module SearchHelper
   def search_placeholder
     "Find dates, sites, references..."
   end
