--- app/helpers/site_types_helper.rb
@@ -1,2 +1,0 @@
-module SiteTypesHelper
-end
