--- app/helpers/sites_helper.rb
@@ -1,7 +1,0 @@
-module SitesHelper
-  include DataTableHelper
-  
-  def wikidata_matches_for(site)
-    Site.wikidata_match_candidates_batch([site])
-  end
-end
