+++ app/helpers/statistics_helper.rb
@@ -1,12 +1,20 @@
-module PagesHelper
+##
+# Helpers methods for project statistics
+module StatisticsHelper
+  ##
+  # Total number of sites (with caching)
   def cached_site_count
     Rails.cache.fetch("stats/site_count", expires_in: 10.minutes) { Site.count }
   end
 
+  ##
+  # Total number of c14s (with caching)
   def cached_c14_count
     Rails.cache.fetch("stats/c14_count", expires_in: 10.minutes) { C14.count }
   end
 
+  ##
+  # Total number of typos (with caching)
   def cached_typo_count
     Rails.cache.fetch("stats/typo_count", expires_in: 10.minutes) { Typo.count }
   end
