--- app/helpers/taxons_helper.rb
@@ -1,2 +1,0 @@
-module TaxonsHelper
-end
