--- app/helpers/typos_helper.rb
@@ -1,3 +1,0 @@
-module TyposHelper
-  include DataTableHelper
-end
