--- app/helpers/xronos_data_helper.rb
@@ -1,2 +1,0 @@
-module XronosDataHelper
-end
