+++ app/models/typo.rb
@@ -19,8 +19,6 @@
 #
 
 class Typo < ApplicationRecord
-  include XronosDataHelper
-  
   acts_as_copy_target # enable CSV exports
 
   validates :name, presence: true
